currency-tracker/parser/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
lxml==5.1.0
flask==3.0.0
schedule==1.2.0
matplotlib==3.8.0
